feat: 불필요한 state제거와 type narrowing
